+++ sqlalchemy_utils/__init__.py
@@ -56,7 +56,7 @@ from .types import (
 )
 
 
-__version__ = '0.23.5'
+__version__ = '0.24.0'
 
 
 __all__ = (