+++ CHANGES.rst
@@ -8,6 +8,7 @@ Here you can see the full list of changes between each SQLAlchemy-Utils release.
 ^^^^^^^^^^^^^^^^^^^
 
 - Added getdotattr
+- Added Path and AttrPath classes
 
 
 0.23.5 (2014-02-15)